requirements.txt
python-dotenv==1.1.1
snowflake==1.8.0